+++ GSC.Data/Entities/CTMS/StudyPlan.cs
@@ -27,7 +27,7 @@ namespace GSC.Data.Entities.CTMS
         public DateTime? ActualStartDate { get; set; }
         public DateTime? ActualEndDate { get; set; }
         public int? CurrencyId { get; set; } // Global Currency
-        public int? TotalCost { get; set; }
+        public decimal? TotalCost { get; set; }
         public Master.Project Project { get; set; }
 
         public Currency Currency { get; set; }
